requirements.txt
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langchain-openai==0.0.5

openai==1.11.0

transformers==4.37.2
python-dotenv==1.0.1